Separate out all errors in metadata service to separate file
