+++ pkg/metadata/property.go
@@ -3,48 +3,10 @@ package metadata
 import (
 	"context"
 
-	"google.golang.org/grpc/codes"
-	"google.golang.org/grpc/status"
-
 	pb "github.com/runmachine-io/runmachine/proto"
 
 	"github.com/runmachine-io/runmachine/pkg/errors"
 	"github.com/runmachine-io/runmachine/pkg/metadata/storage"
-)
-
-var (
-	ErrUnknown = status.Errorf(
-		codes.Unknown,
-		"an unknown error occurred.",
-	)
-	ErrNotFound = status.Errorf(
-		codes.NotFound,
-		"object could not be found.",
-	)
-	ErrPartitionUnknown = status.Errorf(
-		codes.FailedPrecondition,
-		"unknown partition.",
-	)
-	ErrPartitionRequired = status.Errorf(
-		codes.FailedPrecondition,
-		"partition is required.",
-	)
-	ErrObjectTypeRequired = status.Errorf(
-		codes.FailedPrecondition,
-		"object type is required.",
-	)
-	ErrPropertyKeyRequired = status.Errorf(
-		codes.FailedPrecondition,
-		"property key is required.",
-	)
-	ErrSchemaRequired = status.Errorf(
-		codes.FailedPrecondition,
-		"schema is required.",
-	)
-	ErrPropertySchemaObjectRequired = status.Errorf(
-		codes.FailedPrecondition,
-		"property schema object is required.",
-	)
 )
 
 func (s *Server) PropertySchemaDelete(
